challenge2_pitch_visualizer/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
jinja2==3.1.2
nltk>=3.8.0
openai>=1.0.0
textblob==0.17.1
